+++ AplicacionWeb/CafeteriaAPI/Controllers/SalesHistoryController.cs
@@ -1,11 +1,14 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Model.DTO;
+using Model.Models;
 using Model.ViewModel;
 using Service.IServices;
 using Service.Services;
 using System;
 using System.Collections.Generic;
+using System.Security.Claims;
 
 namespace CafeteriaAPI.Controllers
 {
@@ -23,6 +26,7 @@ namespace CafeteriaAPI.Controllers
         }
 
         [HttpGet("GetSalesHistory")]
+        [Authorize]
         public ActionResult<List<SalesHistoryDTO>> GetSalesHistory()
         {
             try
@@ -32,67 +36,91 @@ namespace CafeteriaAPI.Controllers
             }
             catch (Exception ex)
             {
-                _logger.LogError($"Ocurrio un error en GetSalesHistory: {ex.Message}");
+                _logger.LogError($"Ocurrio un error en GetSalesHistory: {ex}");
                 return BadRequest($"{ex.Message}");
             }
         }
 
         [HttpPost("CreateRecord")]
+        [Authorize]
         public ActionResult<SalesHistoryDTO> CreateRecord([FromBody] SalesViewModel record)
         {
             try
             {
-                var response = _salesService.CreateRecord(record);
-                return Ok(response);
+                if (HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Role).Value == "Usuario" || HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Role).Value == "Administrador")
+                {
+                    var response = _salesService.CreateRecord(record);
+                    return Ok(response);
+                }
+                throw new Exception("Necesita registrarse para realizar una compra");
+                
             }
             catch (Exception ex)
             {
-                _logger.LogError($"Ocurrio un error en CreateRecord: {ex.Message}");
+                _logger.LogError($"Ocurrio un error en CreateRecord: {ex}");
                 return BadRequest($"{ex.Message}");
             }
         }
 
         [HttpGet("GetSaleById/{id}")]
+        [Authorize]
         public ActionResult<SalesHistoryDTO> GetSaleById(int id)
         {
             try
             {
-                var response = _salesService.GetSaleById(id);
-                return Ok(response);
+                if (HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Role).Value == "Usuario" || HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Role).Value == "Administrador")
+                {
+                    var response = _salesService.GetSaleById(id);
+                    return Ok(response);
+                }
+                throw new Exception("Necesita registrarse para realizar esta acción");
+                
             }
             catch (Exception ex)
             {
-                _logger.LogError($"Ocurrio un error en GetSaleById: {ex.Message}");
+                _logger.LogError($"Ocurrio un error en GetSaleById: {ex}");
                 return BadRequest($"{ex.Message}");
             }
         }
 
         [HttpPut("ModifySale/{id}")]
+        [Authorize]
         public ActionResult<SalesHistoryDTO> ModifySaleHistory(int id, [FromBody] SalesViewModel record)
         {
             try
             {
-                var response = _salesService.ModifySaleHistory(id, record);
-                return Ok(response);
+                if (HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Role).Value == "Administrador")
+                {
+                    var response = _salesService.ModifySaleHistory(id, record);
+                    return Ok(response);
+                }
+                throw new Exception("No tiene rol Administrador");
+                
             }
             catch (Exception ex)
             {
-                _logger.LogError($"Ocurrio un error en ModifySale: {ex.Message}");
+                _logger.LogError($"Ocurrio un error en ModifySale: {ex}");
                 return BadRequest($"{ex.Message}");
             }
         }
 
         [HttpDelete("DeleteRecord/{id}")]
+        [Authorize]
         public ActionResult<SalesHistoryDTO> DeleteSaleHistory(int id)
         {
             try
             {
-                var response = _salesService.DeleteSaleHistory(id);
-                return Ok(response);
+                if (HttpContext.User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.Role).Value == "Administrador")
+                {
+                    var response = _salesService.DeleteSaleHistory(id);
+                    return Ok(response);
+                }
+                throw new Exception("No tiene rol Administrador");
+                
             }
             catch (Exception ex)
             {
-                _logger.LogError($"Ocurrio un error en DeleteRecord: {ex.Message}");
+                _logger.LogError($"Ocurrio un error en DeleteRecord: {ex}");
                 return BadRequest($"{ex.Message}");
             }
         }
